Agregamos verificacion de token y autenticacion, ademas de nuevas vistas con ejs

+++ src/controllers/posts.controllers.js
@@ -1,10 +1,15 @@
+const { autenticacion } = require("../jwt/autenticacion.js");
 const postsModel = require("../models/posts.model.js");
 const usersModel = require("../models/users.model.js");
 
 class postsControllers {
-  async add(post, user) {
+  async add(token, post, user) {
     return new Promise(async (resolve, reject) => {
       try {
+        const acceso = await autenticacion(token, ["user", "admin"]);
+        if (acceso.mensaje != "acceso permitido") {
+          return reject(acceso.mensaje);
+        }
         const verifyUser = await usersModel.findOne({ username: user });
         if (!verifyUser) {
           return reject("No existe el usuario que esta publicando");
@@ -43,7 +48,11 @@ class postsControllers {
         };
         const datos = await postsModel.create(data); // Creamos el usuario
         if (datos) {
-          return resolve(datos);
+          return resolve({
+            agregado: datos,
+            token: token,
+            username: acceso.username
+          });
         }
         return reject("No se pudo registrar el usuario");
       } catch (error) {
@@ -52,9 +61,13 @@ class postsControllers {
     });
   }
 
-  async update(post, id) {
+  async update(token, post, id) {
     return new Promise(async (resolve, reject) => {
       try {
+        const acceso = await autenticacion(token, ["user", "admin"]);
+        if (acceso.mensaje != "acceso permitido") {
+          return reject(acceso.mensaje);
+        }
         const lastPosts = await postsModel.findById(id); // Validamos que exista el usuario
         console.log(lastPosts);
         if (!lastPosts) {
@@ -85,7 +98,11 @@ class postsControllers {
         };
         const datos = await postsModel.findByIdAndUpdate(id, data); // Creamos el usuario
         if (datos) {
-          return resolve(datos);
+          return resolve({
+            editado: datos,
+            token: token,
+            username: acceso.username
+          });
         }
         return reject("No se pudo registrar el usuario");
       } catch (error) {
@@ -94,16 +111,24 @@ class postsControllers {
     });
   }
 
-  async delete(id) {
+  async delete(token, id) {
     return new Promise(async (resolve, reject) => {
       try {
+        const acceso = await autenticacion(token, ["user", "admin"]);
+        if (acceso.mensaje != "acceso permitido") {
+          return reject(acceso.mensaje);
+        }
         const lastPosts = await postsModel.findById(id); // Validamos que exista la publicacion
         if (!lastPosts) {
           return reject("No existe la publicacion");
         }
         const datos = await postsModel.findByIdAndDelete(id); // Eliminamos la publicacion
         if (datos) {
-          return resolve(datos);
+          return resolve({
+            token: token,
+            eliminado: datos,
+            username: acceso.username
+          });
         }
         return reject("No se pudo eliminar la publicacion");
       } catch (error) {
@@ -112,14 +137,22 @@ class postsControllers {
     });
   }
 
-  async comments(id) {
+  async comments(token, id) {
     return new Promise(async (resolve, reject) => {
       try {
+        const acceso = await autenticacion(token, ["user", "admin"]);
+        if (acceso.mensaje != "acceso permitido") {
+          return reject(acceso.mensaje);
+        }
         const lastPosts = await postsModel.findById(id); // Validamos que exista la publicacion
         if (!lastPosts) {
           return reject("No existe la publicacion");
         }
-        resolve(lastPosts.comments)
+        resolve({
+          token: token,
+          comentarios: lastPosts.comments,
+          username: acceso.username
+        })
       } catch (error) {
         return reject(error);
       }
